requirements.txt
# Core Nextcord
nextcord>=2.0.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
aiohttp>=3.9.0

# System monitoring
psutil>=5.9.0

# PostgreSQL database
asyncpg==0.29.0

# Optional but recommended for Windows color support
colorama>=0.4.6 ; platform_system=="Windows"

# Development dependencies (optional)
# pip install -r requirements-dev.txt
# black>=23.0.0
# flake8>=6.0.0
# pytest>=7.0.0
# pytest-asyncio>=0.21.0